* commit '99898de902c36687be126b33ee8a4858d26871ac':
  Fix NdefRecord byte-stream constructor.

+++ core/jni/android_nfc_NdefRecord.cpp
@@ -149,7 +149,7 @@ static jint android_nfc_NdefRecord_parseNdefRecord(JNIEnv *e, jobject o,
 
     /* Set flags field */
     mFlags = e->GetFieldID(record_cls, "mFlags", "B");
-    e->SetIntField(o, mFlags, record.Flags);
+    e->SetByteField(o, mFlags, record.Flags);
 
     ret = 0;
